refactor(issues): merge / split 走 web→api role 中转（对齐现网架构）
PR 此前新增的 MergeIssueResource / SplitIssueResource 直接在 web role 执行写操作，
违反项目既有惯例——现有 10 个状态机方法（resolve / archive / reopen 等）全部走
`api.issue.<method>` → KernelAPIResource → kernel_api/views/v4/issue.py 的 api role
Resource 模式。web role 缺 REDIS_*_CONF，cache 操作会被静默吞，30min TTL 才自然
恢复，cache 与 ES 状态可能错路由新告警到已变更 member 上。

本提交把 merge / split 迁到 api role：

- api/issue/default.py 新增 MergeResource / SplitResource（KernelAPIResource）
- kernel_api/views/v4/issue.py 新增同名 Resource 承载完整校验 + 写关系 + ES 活动
  + cache invalidate；operator 提到 RequestSerializer（参考 ResolveResource）
- packages/fta_web/issue/resources.py 的 MergeIssueResource / SplitIssueResource
  改为薄壳：仅校验入参 + 调 api.issue.merge / split（operator 走 get_request_username）
- 删 fta_web 侧 _invalidate_active_members_cache（不再被任何 web 端 Resource 调用）

效果：merge / split 的 cache 操作在 api role 真生效，与 resolve / archive 同链路。
